requirements.txt
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
idna==2.7
oauthlib==2.1.0
praw==5.4.0
prawcore==0.14.0
PySocks==1.6.8
python-twitter==3.4.2
requests==2.19.0
requests-oauthlib==1.0.0
six==1.11.0
SQLAlchemy==1.2.8
SQLAlchemy-Utils==0.33.3
update-checker==0.16
urllib3==1.23
